requirements.txt
pyautogui>=0.9
pyclick>=0.0.2
pynput>=1.7
ttkbootstrap>=1.10
Pillow>=10.0
numpy>=1.26
opencv-python>=4.10
mss>=9.0
